solution/backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Anthropic
anthropic==0.40.0

# Telegram
python-telegram-bot==21.9

# Scheduler (re-engagement jobs)
apscheduler==3.10.4

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utilities
httpx==0.28.1
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
